notemap-project-main
This note map website project is a web application created using HTML, CSS, JavaScript, and an API. The project allows users to organize their notes on a map and store them in either a local or remote database. This user-friendly website is designed for visually organizing notes and tracking their locations.